% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{executar_projeto_florestal}
\alias{executar_projeto_florestal}
\title{Main Forest Project Execution Function}
\usage{
executar_projeto_florestal(
  arq_csv,
  arq_rds,
  pasta_saida = "Resultados_Projeto",
  area_total = 18,
  anos_simulacao = 18,
  gerar_animacao = FALSE
)
}
\arguments{
\item{arq_csv}{Character. Path to the forest inventory CSV.}

\item{arq_rds}{Character. Path to the trained model RDS.}

\item{pasta_saida}{Character. Output folder name.}

\item{area_total}{Numeric. Total area in hectares.}

\item{anos_simulacao}{Numeric. Number of years to simulate.}

\item{gerar_animacao}{Logical. Whether to generate the 3D GIF.}
}
\description{
Main Forest Project Execution Function
}
