% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_corte.R
\name{selecionar_arvores_gui}
\alias{selecionar_arvores_gui}
\title{GUI for Tree Selection (CONAMA 406/2009 Logging Rules)}
\usage{
selecionar_arvores_gui(df, area_total = 18)
}
\description{
GUI for Tree Selection (CONAMA 406/2009 Logging Rules)
}
