% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformar_para_grid}
\alias{transformar_para_grid}
\title{Transforms local plot coordinates into a global spatial grid}
\usage{
transformar_para_grid(df_resultado)
}
\description{
Transforms local plot coordinates into a global spatial grid
}
